+++ odoo/local-src/specific_zetes/__init__.py
@@ -5,3 +5,4 @@ from . import constants
 from . import models
 from . import tools
 from . import controllers
+from . import wizard